1. Added md_frameheader_description method to return a quick cstring description of a frameheader.
2. Added public constant arrays that map the following values of enums
to their string names. Potentially in the future make these arrays
private and create public method interfaces to them:

* MPEGAudioVersionNames
* MPEGLayerNames
* MPEGChannelModeNames

3. Added md_string_appendformat which appends a format through use of
vsprintf. Note though that to use vsprintf, I needed to have a cstring
buffer, but I dont know how big it has to be so I just choose a medium
size power of 2 which should be large enough for most format strings
(Ill cross that bridge when I get there). The current value is 5196.